General edits: parts I-IV
General edits made to Parts I-IV. Specific changes include: (a) I changed all text to be consistently past tense, (b) because the platform used is noted at the start of each section, I have removed in-text comments at the step level stating 'in R' or 'in ArcPro'. To make the text more concise, and (c) I adjusted the text such that, for commands run in R, the package is reflected in the code presented, e.g., 'sf::st_is_valid()'.